Zombie working
For the most part the zombies are working, there is a slight delay on the 3rd hit. But zombie's move when hit, with animation and stay put when dead.

Working on some fine tuning. then will implement a variance to make the zombie run after the first hit increasing its speed.