Flying Zombie commit
Used Fisher-Yates Shuffle to make random points. There is a patrolPoints gameObject that is disabled. It has 9 cube objects.  These act as a 3x3 array of patrol points. The transform array will have the enemy fly to each of these points. So in theory we can move them anywhere on the map and it will patrol the area.