Added 3 more weapon types/ Modified PlayerController to be able to adapt and accept these weapons and set ADS and Weapon position per Weapon Type
Modified Player Weapon Controller to accept a new weapon holder.
Everything that was inside the old gunModel like barrel tip, muzzle flash, ads ect now live inside each weapon.
Each weapon now has their own ADS and hip position, barrelTip, muzzle flash, Shell Ejector ect.
Modified and added a flag to when a player grabs a pistol it will disable the other arm that is not in used.
Move Prefabs to the prefabs folder.
Updated player Prefab.
Created a scop with an UI canvas to add a redDot
Cleaned Up Folder Structure
Added an option in the gunStats to include a sprite icon for Hotbar UI
Added a bool to check if its a pistol