Added an option to randomize ammo on weapon Pickups, and weapon pick up now gives ammo instead of grabbing another weapon
Commit contains changes to weapon assets to include the new code addition to set a weapon id by doing this weapon id this gives us the ability when picking up a weapon if the player already owns the weapon it stops the player from re adding a weapon to inventory and playing the gun take out animation. The new addition allows the player to pick up a weapon and the ammo will be added to current weapon.
Code now check for has gun(any gun) and plays necessary animations
while gun(weapon id) checks if player has a specific weapon with same id, if same id it will add the ammo to current weapon.